move web api requests to separate class
